move lookupTaskBin to executor_linux, for os dependency clarity
