+++ drivers/shared/executor/executor_linux_test.go
@@ -161,6 +161,41 @@ ld.so.conf.d/`
 	}, func(err error) { t.Error(err) })
 }
 
+func TestUniversalExecutor_LookupTaskBin(t *testing.T) {
+	t.Parallel()
+	require := require.New(t)
+
+	// Create a temp dir
+	tmpDir, err := ioutil.TempDir("", "")
+	require.Nil(err)
+	defer os.Remove(tmpDir)
+
+	// Make a foo subdir
+	os.MkdirAll(filepath.Join(tmpDir, "foo"), 0700)
+
+	// Write a file under foo
+	filePath := filepath.Join(tmpDir, "foo", "tmp.txt")
+	err = ioutil.WriteFile(filePath, []byte{1, 2}, os.ModeAppend)
+	require.NoError(err)
+
+	// Lookout with an absolute path to the binary
+	_, err = lookupTaskBin(tmpDir, "/foo/tmp.txt")
+	require.NoError(err)
+
+	// Write a file under local subdir
+	os.MkdirAll(filepath.Join(tmpDir, "local"), 0700)
+	filePath2 := filepath.Join(tmpDir, "local", "tmp.txt")
+	ioutil.WriteFile(filePath2, []byte{1, 2}, os.ModeAppend)
+
+	// Lookup with file name, should find the one we wrote above
+	_, err = lookupTaskBin(tmpDir, "tmp.txt")
+	require.NoError(err)
+
+	// Lookup a host absolute path
+	_, err = lookupTaskBin(tmpDir, "/bin/sh")
+	require.Error(err)
+}
+
 // Exec Launch looks for the binary only inside the chroot
 func TestExecutor_EscapeContainer(t *testing.T) {
 	t.Parallel()
