+++ drivers/shared/executor/executor_test.go
@@ -452,41 +452,6 @@ func TestUniversalExecutor_LookupPath(t *testing.T) {
 	require.NoError(err)
 }
 
-func TestUniversalExecutor_LookupTaskBin(t *testing.T) {
-	t.Parallel()
-	require := require.New(t)
-
-	// Create a temp dir
-	tmpDir, err := ioutil.TempDir("", "")
-	require.Nil(err)
-	defer os.Remove(tmpDir)
-
-	// Make a foo subdir
-	os.MkdirAll(filepath.Join(tmpDir, "foo"), 0700)
-
-	// Write a file under foo
-	filePath := filepath.Join(tmpDir, "foo", "tmp.txt")
-	err = ioutil.WriteFile(filePath, []byte{1, 2}, os.ModeAppend)
-	require.NoError(err)
-
-	// Lookout with an absolute path to the binary
-	_, err = lookupTaskBin(tmpDir, "/foo/tmp.txt")
-	require.NoError(err)
-
-	// Write a file under local subdir
-	os.MkdirAll(filepath.Join(tmpDir, "local"), 0700)
-	filePath2 := filepath.Join(tmpDir, "local", "tmp.txt")
-	ioutil.WriteFile(filePath2, []byte{1, 2}, os.ModeAppend)
-
-	// Lookup with file name, should find the one we wrote above
-	_, err = lookupTaskBin(tmpDir, "tmp.txt")
-	require.NoError(err)
-
-	// Lookup a host absolute path
-	_, err = lookupTaskBin(tmpDir, "/bin/sh")
-	require.Error(err)
-}
-
 // setupRoootfs setups the rootfs for libcontainer executor
 // It uses busybox to make some binaries available - somewhat cheaper
 // than mounting the underlying host filesystem
